refactor: 상수파일명 database를 data로 변경
